Autocopy of .dist files when building
When building the project, *.dist files are copied at the same
place _unless_ there already exists a file with that name.

This allows us to check .dist files into the respository, while
avoiding to check in configuration files (that will oftentimes
change, and/or contain local settings that do not apply to all
the developers).