Added hysteresis factor to avoid target switching
The straight line distance from an agent to a target is used when
computing the utility of an agent/target pair. However, the agent's
pathing is hardly ever a straight line. Thus, sometimes an agent
will keep switching from target A to target B just because the path
from its location to A goes through a point where the agent is
(straight-line) closer to B than A, and viceversa.

The hysteresis factor introduced in this patch prevent such behavior,
by introducing a multiplier that increases the utility of keeping the
current target.