Implement NCCCs, byte counting and cleanup DSA
Implemented byte and NCCCs counting (the algorithms themselves are
responsible of reporting these values. The simulator just collects them
and saves them to the results file.

Also cleaned up some DSA clustering features that have no sense because
the communication range limitation is not considered in the current
state of the simulator.