Submit assignments directly to fire agents
Up until now, assignment messages and agent locations where transmitted
between the CenterAgent and the FireAgents using the roborescue's
communication layer. The problem is that such layer introduces a huge
latency (messages are not received until the following simulation step).

This may or may not be a good constraint to have. However, it makes no
sense at all to impose this contraint when the agents themselves are
supposed to be communicating using a much lower latency network. Hence,
this commit avoids using the roborescue's communication channel, and
assignments are transmitted directly to the FireAgents.

In turn, at each simulation step the FireAgents wait until they receive
that step's assignment before performing any actions.